+++ covalent_dispatcher/_dal/importers/lattice.py
@@ -24,14 +24,12 @@ from sqlalchemy.orm import Session
 
 from covalent._shared_files.config import get_config
 from covalent._shared_files.schemas.lattice import (
-    LATTICE_CALL_AFTER_FILENAME,
-    LATTICE_CALL_BEFORE_FILENAME,
     LATTICE_COVA_IMPORTS_FILENAME,
-    LATTICE_DEPS_FILENAME,
     LATTICE_DOCSTRING_FILENAME,
     LATTICE_ERROR_FILENAME,
     LATTICE_FUNCTION_FILENAME,
     LATTICE_FUNCTION_STRING_FILENAME,
+    LATTICE_HOOKS_FILENAME,
     LATTICE_INPUTS_FILENAME,
     LATTICE_LATTICE_IMPORTS_FILENAME,
     LATTICE_NAMED_ARGS_FILENAME,
@@ -76,9 +74,7 @@ def _get_lattice_meta(lat: LatticeSchema, storage_path) -> dict:
         "named_args_filename": LATTICE_NAMED_ARGS_FILENAME,
         "named_kwargs_filename": LATTICE_NAMED_KWARGS_FILENAME,
         "results_filename": LATTICE_RESULTS_FILENAME,
-        "deps_filename": LATTICE_DEPS_FILENAME,
-        "call_before_filename": LATTICE_CALL_BEFORE_FILENAME,
-        "call_after_filename": LATTICE_CALL_AFTER_FILENAME,
+        "hooks_filename": LATTICE_HOOKS_FILENAME,
         "cova_imports_filename": LATTICE_COVA_IMPORTS_FILENAME,
         "lattice_imports_filename": LATTICE_LATTICE_IMPORTS_FILENAME,
     }